ovdserver/backends/OWLv2/requirements.txt
transformers
Pillow
torch
opencv-python
opencv-python-headless
fastapi
uvicorn